+++ backend-lib/src/main/java/cash/z/ecc/android/sdk/internal/Backend.kt
@@ -10,6 +10,7 @@ import cash.z.ecc.android.sdk.internal.model.JniSubtreeRoot
 import cash.z.ecc.android.sdk.internal.model.JniTransactionDataRequest
 import cash.z.ecc.android.sdk.internal.model.JniWalletSummary
 import cash.z.ecc.android.sdk.internal.model.ProposalUnsafe
+import cash.z.ecc.android.sdk.model.UnifiedAddressRequest
 
 /**
  * Contract defining the exposed capabilities of the Rust backend.
@@ -194,6 +195,12 @@ interface Backend {
 
     @Throws(RuntimeException::class)
     suspend fun getCurrentAddress(accountUuid: ByteArray): String
+
+    @Throws(RuntimeException::class)
+    suspend fun getNextAvailableAddress(
+        accountUuid: ByteArray,
+        receiverFlags: Int
+    ): String
 
     fun getTransparentReceiver(ua: String): String?