+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/TypesafeBackend.kt
@@ -24,6 +24,7 @@ import cash.z.ecc.android.sdk.model.UnifiedFullViewingKey
 import cash.z.ecc.android.sdk.model.UnifiedSpendingKey
 import cash.z.ecc.android.sdk.model.Zatoshi
 import cash.z.ecc.android.sdk.model.ZcashNetwork
+import cash.z.ecc.android.sdk.model.UnifiedAddressRequest
 
 @Suppress("TooManyFunctions")
 internal interface TypesafeBackend {
@@ -131,6 +132,12 @@ internal interface TypesafeBackend {
 
     @Throws(RustLayerException.GetAddressException::class)
     suspend fun getCurrentAddress(account: Account): String
+
+    @Throws(RustLayerException.GetAddressException::class)
+    suspend fun getNextAvailableAddress(
+        account: Account,
+        request: UnifiedAddressRequest
+    ): String
 
     suspend fun listTransparentReceivers(account: Account): List<String>